pyproject.toml
[project]
name = "fire-smoke-cv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.1",
    "kagglehub>=1.0.1",
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "python-multipart>=0.0.27",
    "pyyaml>=6.0.3",
    "tqdm>=4.67.3",
    "ultralytics>=8.4.43",
    "uvicorn>=0.46.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "ruff>=0.15.12",
]
